moves DimcatResource from base to data/base
