+++ src/dimcat/__init__.py
@@ -17,7 +17,7 @@ from .analyzer import (
     PitchClassVectors,
     TPCrange,
 )
-from .data import Dataset
+from .dataset import Dataset
 from .filter import HasCadenceAnnotationsFilter, IsAnnotatedFilter
 from .grouper import CorpusGrouper, ModeGrouper, PieceGrouper, YearGrouper
 from .pipeline import Pipeline
