+++ src/dimcat/features/annotations.py
@@ -1,21 +1,9 @@
-from dataclasses import dataclass
-
-from dimcat.features.base import Feature, FeatureConfig, FeatureName
-
-
-@dataclass(frozen=True)
-class AnnotationsConfig(FeatureConfig):
-    _configured_class = FeatureName.Annotations
+from dimcat.features.base import Feature
 
 
 class Annotations(Feature):
-    _config_type = AnnotationsConfig
-
-
-@dataclass(frozen=True)
-class KeyAnnotationsConfig(AnnotationsConfig):
-    _configured_class = FeatureName.KeyAnnotations
+    pass
 
 
 class KeyAnnotations(Annotations):
-    _config_type = KeyAnnotationsConfig
+    pass
